chatbot/requirements.txt
aiogram
psycopg2-binary
python-dotenv
sqlalchemy
